+++ amaru/float128.c
@@ -30,7 +30,7 @@ amaru_float128_to_ieee128(float128_t const value) {
   amaru128_fields_t binary;
   binary.mantissa = amaru_lsb(bits, mantissa_size);
   bits >>= mantissa_size;
-  binary.exponent = amaru_lsb(bits, exponent_size);
+  binary.exponent = (int32_t) amaru_lsb(bits, exponent_size);
 
   return binary;
 }
